Minor fixes for VC compatibility, boost::regex dependency almost optional (#321)
* Added another test case

* Add GetOpposingEdgeId method.

* Fix cut and paste fail.

* Add a flag for a directed edge that ends at a dead-end node.

* added zone abbrev to the get_duration output and removed the tz offset.

* updated tests with the zone abbrev

* Set maximum values for trip Id and headsign offset.

* added offset back and tz to the end....thor will remove tz and send back via protocol buffer.

* Added GetOffsetForHeading method that will adjust distance based on road class

* Added 'use' in the GetOffsetForHeading calculation

* Add max_up_slope and max_down_slope (and set methods). Uses 1 degree
precision up to 16 degrees and 4 degrees afterwards (up to max of 76 degrees).

* Moved from thor. Removed dependency on sif EdgeLabels, now accepts
a function to get the sortcost from a specified label index within
the empty_overflow method.

* strenghting tile null checks (#275)

* Use a lambda for getting the next label from a bucket.

* Move to a more general class name: DoubleBucketQueue

* Replace with new name: DoubleBucketQueue

* Missed this one on last commit.

* added onestop maps.

* no need to do a binary search for the departures just insert them into a map.

* my bad.

* changed tile_line_pair to tile_index_pair.  Added GetStopOneStops.  added stop_one_stops for including and excluding stops.

* changed tile_line_pair to tile_index_pair.  Added GetStopOneStops.  added stop_one_stops for including and excluding stops.  Added logic to populated the map.

* Header path fix

* Fix Visual C++ compilation error

* Fix Visual C++ compilation error

* Include 'config.h' for platform-specific workarounds

* Remove unnecessary boost filesystem dependency

* Replace boost::regex with wrapper that can be configured either as boost::regex (default) or std::regex (enabled is USE_STD_REGEX preprocessor variable is set)

* Introduced GraphTileStorage interface that handles all tile related discovery and IO, filesystem based storage handling is implemented in GraphTileFsStorage

* Various fixes

* Replace boost::regex dependency with configurable (USE_STD_REGEX) std::regex/boost::regex dependency

* more obvious local include

+++ src/baldr/directededge.cc
@@ -579,8 +579,8 @@ void DirectedEdge::set_max_up_slope(const float slope) {
 // Gets the maximum downward slope. Uses 1 degree precision for slopes to
 // -8 degrees, and 4 degree precision afterwards (up to a max of -76 degs).
 int DirectedEdge::max_down_slope() const {
-  return ((max_down_slope_ & 0x10) == 0) ? -max_down_slope_ :
-          -(16 + ((max_down_slope_ & 0xf) * 4));
+  return ((max_down_slope_ & 0x10) == 0) ? -static_cast<int>(max_down_slope_) :
+          -static_cast<int>(16 + ((max_down_slope_ & 0xf) * 4));
 }
 
 // Sets the maximum downward slope.
